requirements.txt
beautifulsoup4==4.11.1
humanize==4.1.0
metadata_parser==0.10.5
Pillow==9.1.0
requests==2.27.1
tqdm==4.64.0
tweepy==4.8.0
zeroc-ice==3.7.10.1
